Add special handling for errors resulting from tesseract shell command. If the tesseract shell command fails (tesseract is not installed, command-line options are malformed, etc.) then we can raise a slightly more informative error.
We do some extra checking (e.g. for non-nil $?) to hopefully be
compatible with non-POSIX systems like Windows, which may not behave the
same way w.r.t ruby's Process::Status.

Some sources state that POSIX shells in general should use code 127 in
this case, though I was not able to find a canonical source and
Wikipedia indicates this may not be true for all shells.